SLES15 STIG : products/sle15/profiles/stig.profile: Fix value for SLES-15-030590
SLES-15-030590 is saying that 'syslog' is a valid value but the check is looking for 'single'
which is not even the value shown in the rule (the rule is asking to check for 'syslog' and
ask to set 'halt'. Since the check is talking of 'syslog', this value has been choosen.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>